feat: remove Alt-F5 from GNOME unmaximize keybinding
The exposed GNOME keybinding for GNOME is '<Super>Down', and it's all a
user sees in the GNOME Settings. However, '<Alt>F5' is also there as a
hidden default. This is confusing enough, but the real trouble is this
shortcut conflicts with VSCode's default "next diff" shortcut. As this
hidden shortcut seems somewhat deprecated, I'm removing it here.